Adding Files for initial commit
You project will need to have the following features:
Create a new Empty Application Project

Create a variable using the float data type. Cast the float to an int
and using NSLog, output both the initial float value as well as the int
value.

Perform an AND, OR comparison. Use float, int and BOOL types. BOOL
values should be YES or NO and written in all caps.